@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
-import { TranslateModule } from '@ngx-translate/core';
 import { IonicPageModule } from 'ionic-angular';
+import { TranslateModule } from '@ngx-translate/core';
 
 import { SettingsPage } from './settings';
 
@@ -10,10 +10,7 @@ import { SettingsPage } from './settings';
   ],
   imports: [
     IonicPageModule.forChild(SettingsPage),
-    TranslateModule.forChild()
+    TranslateModule.forChild(),
   ],
-  exports: [
-    SettingsPage
-  ]
 })
-export class SettingsPageModule { }
+export class SettingsPageModule {}
