@@ -1,95 +1,49 @@
-import { Component } from '@angular/core';
-import { FormBuilder, FormGroup } from '@angular/forms';
-import { TranslateService } from '@ngx-translate/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
+import { Component, Inject } from '@angular/core';
+import { IonicPage, ToastController } from 'ionic-angular';
+import * as Clipboard from 'clipboard';
 
-import { Settings } from '../../providers/providers';
+import { AuthenticationProvider } from './../../providers/authentication/authentication';
+import { UserProvider } from './../../providers/api-services/users';
+import { EnvVariables, IEnvVariables } from '../../modules/environment-variables/environment-variables.token';
 
-/**
- * The Settings page is a simple form that syncs with a Settings provider
- * to enable the user to customize settings for the app.
- *
- */
 @IonicPage()
 @Component({
   selector: 'page-settings',
-  templateUrl: 'settings.html'
+  templateUrl: 'settings.html',
 })
 export class SettingsPage {
-  // Our local settings object
-  options: any;
 
-  settingsReady = false;
+  private clipboard;
 
-  form: FormGroup;
+  public feedUrl: string;
 
-  profileSettings = {
-    page: 'profile',
-    pageTitleKey: 'SETTINGS_PAGE_PROFILE'
-  };
-
-  page: string = 'main';
-  pageTitleKey: string = 'SETTINGS_TITLE';
-  pageTitle: string;
-
-  subSettings: any = SettingsPage;
-
-  constructor(public navCtrl: NavController,
-    public settings: Settings,
-    public formBuilder: FormBuilder,
-    public navParams: NavParams,
-    public translate: TranslateService) {
+  constructor(
+    private userProvider: UserProvider,
+    @Inject(EnvVariables) private envVariables: IEnvVariables,
+    private toastController: ToastController,
+    private authenticationProvider: AuthenticationProvider,
+  ) {
   }
 
-  _buildForm() {
-    let group: any = {
-      option1: [this.options.option1],
-      option2: [this.options.option2],
-      option3: [this.options.option3]
-    };
-
-    switch (this.page) {
-      case 'main':
-        break;
-      case 'profile':
-        group = {
-          option4: [this.options.option4]
-        };
-        break;
-    }
-    this.form = this.formBuilder.group(group);
-
-    // Watch the form for changes, and
-    this.form.valueChanges.subscribe((v) => {
-      this.settings.merge(this.form.value);
+  public ionViewDidEnter(): void {
+    this.clipboard = new Clipboard('#feed-btn', {
+      text: () => this.feedUrl
     });
+    this.clipboard.on('success', () => this.toastController.create({ message: 'Feed kopiert', duration: 3000, position: 'bottom' }).present());
+    this.clipboard.on('error', () => this.toastController.create({ message: 'Feed konnte nicht kopiert werden', duration: 3000, position: 'bottom' }).present());
+
+    this.userProvider.getMe()
+      .subscribe(user => {
+        this.feedUrl = this.envVariables.apiEndpoint + 'feed/' + user.calenderToken;
+      });
   }
 
-  ionViewDidLoad() {
-    // Build an empty form for the template to render
-    this.form = this.formBuilder.group({});
+  public ionViewWillLeave(): void {
+    this.clipboard.destroy();
   }
 
-  ionViewWillEnter() {
-    // Build an empty form for the template to render
-    this.form = this.formBuilder.group({});
-
-    this.page = this.navParams.get('page') || this.page;
-    this.pageTitleKey = this.navParams.get('pageTitleKey') || this.pageTitleKey;
-
-    this.translate.get(this.pageTitleKey).subscribe((res) => {
-      this.pageTitle = res;
-    })
-
-    this.settings.load().then(() => {
-      this.settingsReady = true;
-      this.options = this.settings.allSettings;
-
-      this._buildForm();
-    });
+  public logout(): void {
+    this.authenticationProvider.logout();
   }
 
-  ngOnChanges() {
-    console.log('Ng All Changes');
-  }
 }
